requirements.txt
fastapi==0.115.6
pillow==11.0.0
python-multipart==0.0.20
tensorflow==2.18.0
uvicorn==0.34.0
